Added example component that connects to DB
